feat(tetris): enhance game mechanics and dynamic import
